Remove `raise ValueError(f"self.config should be None")`
Sometimes self.config on old line 69 is still populated and we don't understand why. Maybe the contextmanager never executes the exit statement after yield? Although we should find the core issue and solve that, for now let's remove the validation because self.config is overwritten anyway, and the ValueError causes apps to be broken.